requirements.txt
pyautogen>=0.2.0
openai>=1.0.0
streamlit>=1.28.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
markdown>=3.5.0
jinja2>=3.1.0
requests>=2.31.0
typing-extensions>=4.8.0 
black
autogen
autogen-agentchat